+++ lib/gl-help.js
